+++ game/rendering/ModelRenderer.h
@@ -44,7 +44,7 @@ class ModelRenderer {
 
     void updateModel(
         uint64_t id
-      , size_t frameIdx
+      , uint32_t frameIdx
       , const glm::mat4& mvp
       , const glm::vec3& cameraPos
       , const glm::mat4& normal
